Always return 404 for Failures in Zarr Streaming
